Crate nut

This is NUT!!! 🤣

This is my little experiment in code generation. Take a model and turn it into code. Since my favorite, and about the only language I can work in, is Rust, I’m going to generate Rust. At least that’s the plan. Don’t shoot me if I generate some Typescript along the way. It could happen.

Module Structure

I’m doing the code generation in stages. I sort of have a general idea of how it will work. Nothing quite coherent though. So I’m nudging things in the general direction I’m feeling at the time. When it feels like a good place to stop, which has generally been one simple transformation, then I start a new module.

ooa_0

I started with a handful of [structs and enums][ooa_0] and a binary that created instances of said enums and spit out JSON. I got bored creating instances of them by hand. So I built Cuckoo, and together they evolved.

I got to the point that Cuckoo is pretty stable, and it seemed that [ooa_0] would be frozen. That hasn’t turned out to be exactly true, but the Cuckoo facing stuff has been mostly stable.

So it seemed like a good time to start working on [ooa_1].

ooa_1

This is a nascent work in progress. The general direction it seems to be taking is changing the Javascript friendly JSON into something more abstract. Well, maybe it’s not changing the level of abstraction. It’s rendering a homogeneous way to navigate relationships. I’m thinking that this is going to be driven by relationships. If they are in fact functors, then that would make sense, I think.

We’ll see how it goes.

ooa_2

It went ok. The problem we are solving now is adding relationship pointers to objects. That is, for each relationship connected to an object, we want to make it available to the object itself. We didn’t do this last time because we didn’t have updated relationships during extrusion. Now we do!

ooa_1 & 2 redo

I assume this is going to get cleaned out sometime, and this will disappear. In the mean time, notes! It turns out that ooa_2 was a mistake. I’m going the total other direction – normalizing it all. So I’m redoing ooa_1 basically.

model

So the redo went really well. I think everything is normalized now. There are some macros for relationship navigation. All in all, it’s pretty slick.

Now I need to import the model domain side of things from Cuckoo. It’s for Nutter, the new UI.

drawing

This is generated. It’s the domain for storing drawing geometry. The program that performs the generation is bin/drawing/gen_code.rs.

Modules

codegen

Code Generation Doings

domain

drawing

The Drawing Domain

sarzak

Sarzak OOA of OOA

Macros

emit

Emit a String without a line break added

emitln

Emit a String with a line break added

function

This macro prints the name of the function in which it’s called.

get_a_across_r1

Macro to traverse Object ➡ Attribute, via R1

get_a_across_r2

Macro to traverse Type ➡ Attribute, via R2

get_a_across_r3

Macro to traverse Edge ➡ Anchor, via R3

get_a_across_r4

Macro to traverse Point ➡ Anchor, via R4

get_a_across_r5

Macro to traverse Point ➡ Anchor, via R5

get_a_across_r7

Macro to traverse BinaryUI ➡ Anchor, via R7

get_a_across_r8

Macro to traverse BinaryUI ➡ Anchor, via R8

get_a_across_r9

Macro to traverse IsaUI ➡ Anchor, via R9

get_a_across_r10

Macro to traverse IsaUI ➡ Anchor, via R10

get_bui_across_r7

Macro to traverse Anchor ➡ BinaryUI, via R7

get_bui_across_r8

Macro to traverse Anchor ➡ BinaryUI, via R8

get_bui_across_r12

Macro to traverse Binary ➡ BinaryUI, via R12

get_card_across_r8

Macro to traverse Referent ➡ Cardinality, via R8

get_card_across_r9

Macro to traverse Referrer ➡ Cardinality, via R9

get_cond_across_r11

Macro to traverse Referrer ➡ Conditionality, via R11

get_cond_across_r12

Macro to traverse Referent ➡ Conditionality, via R12

get_e_across_r3

Macro to traverse Anchor ➡ Edge, via R3

get_e_across_r14

Macro to traverse ObjectUI ➡ Edge, via R14

get_e_across_r19

Macro to traverse Object ➡ Event, via R19

get_e_across_r20

Macro to traverse State ➡ Event, via R20

get_iui_across_r9

Macro to traverse Anchor ➡ IsaUI, via R9

get_iui_across_r10

Macro to traverse Anchor ➡ IsaUI, via R10

get_iui_across_r11

Macro to traverse Isa ➡ IsaUI, via R11

get_many_anch_across_r10

❌ Macro to traverse IsaUI ➡ Anchor, via R10

get_many_edg_across_r14

❌ Macro to traverse ObjectUI ➡ Edge, via R14

get_obj_across_r1

Macro to traverse Attribute ➡ Object, via R1

get_obj_across_r14

Macro to traverse Supertype ➡ Object, via R14

get_obj_across_r15

Macro to traverse Subtype ➡ Object, via R15

get_obj_across_r16

Macro to traverse Referent ➡ Object, via R16

get_obj_across_r17

Macro to traverse Referrer ➡ Object, via R17

get_obj_across_r18

Macro to traverse State ➡ Object, via R18

get_obj_across_r19

Macro to traverse Event ➡ Object, via R19

get_one_anch_across_r7

Macro to traverse BinaryUI ➡ Anchor, via R7

get_one_anch_across_r8

Macro to traverse BinaryUI ➡ Anchor, via R8

get_one_anch_across_r9

Macro to traverse IsaUI ➡ Anchor, via R9

get_one_pnt_across_r4

Macro to traverse Anchor ➡ Point, via R4

get_one_pnt_across_r5

Macro to traverse Anchor ➡ Point, via R5

get_one_pnt_across_r13

Macro to traverse ObjectUI ➡ Point, via R13

get_one_r_bin_across_r12

Macro to traverse BinaryUI ➡ Binary, via R12

get_oui_across_r1

Macro to traverse Object ➡ ObjectUI, via R1

get_oui_across_r13

Macro to traverse Point ➡ ObjectUI, via R13

get_oui_across_r14

Macro to traverse Edge ➡ ObjectUI, via R14

get_p_across_r4

Macro to traverse Anchor ➡ Point, via R4

get_p_across_r5

Macro to traverse Anchor ➡ Point, via R5

get_p_across_r13

Macro to traverse ObjectUI ➡ Point, via R13

get_r_bin_across_r5

Macro to traverse Referent ➡ Binary, via R5

get_r_bin_across_r6

Macro to traverse Referrer ➡ Binary, via R6

get_r_bin_across_r12

Macro to traverse BinaryUI ➡ Binary, via R12

get_r_from_across_r6

Macro to traverse Binary ➡ Referrer, via R6

get_r_from_across_r9

Macro to traverse Cardinality ➡ Referrer, via R9

get_r_from_across_r11

Macro to traverse Conditionality ➡ Referrer, via R11

get_r_from_across_r17

Macro to traverse Object ➡ Referrer, via R17

get_r_isa_across_r11

Macro to traverse IsaUI ➡ Isa, via R11

get_r_sub_across_r15

Macro to traverse Object ➡ Subtype, via R15

get_r_sup_across_r14

Macro to traverse Object ➡ Supertype, via R14

get_r_to_across_r5

Macro to traverse Binary ➡ Referent, via R5

get_r_to_across_r8

Macro to traverse Cardinality ➡ Referent, via R8

get_r_to_across_r12

Macro to traverse Conditionality ➡ Referent, via R12

get_r_to_across_r16

Macro to traverse Object ➡ Referent, via R16

get_referent

Macro to get the Referent side of a Binary relationship

get_referrer

Macro to get the Referrer side of a Binary relationship

get_s_across_r18

Macro to traverse Object ➡ State, via R18

get_s_across_r20

Macro to traverse Event ➡ State, via R20

get_subtypes

Macro to get the Supertype from an Isa relationship

get_supertype

Macro to get the Supertype from an Isa relationship

get_t_across_r2

Macro to traverse Attribute ➡ Type, via R2

maybe_get_one_edg_across_r3

Macro to traverse Anchor ➡ Edge, via R3(c)

Structs

Object_v0

In general an object is a memory representation of a set of collected attributes/fields/members (we call them Attributes), and functions. We are currently ignoring functions.

Schema_v0

ooa_0 Schema

Schema_v1

OOA_1 Schema

Enums

ObjectType

“Type erase” our objects

Relationship_V1

An association between two or more objects

Relationship_v0

Constants

UUID_NS_v0

Traits

ReadCuckooModel

Read the Cuckoo

ReadSchema_v0

ReadSchema_v1

WriteSchema_v0

WriteSchema_v1

Functions

extract_ooa2