Module nut::codegen

Code Generation Doings

This module is mostly re-exports of other interesting things for code generation, pulled together from other modules. This will always contain the latest things, which should make use of this crate easier, I hope.

Re-exports

pub use config::Config;

pub use config::ConfigValue;

pub use config::ImportedObject;

pub use config::SingletonObject;

pub use context::CachingContext;

pub use context::Context;

pub use types::Attr;

pub use types::Field;

pub use types::Ref;

pub use emit;

pub use emitln;

pub use function;

pub use crate::ooa_2::get_referent;

pub use crate::ooa_2::get_referrer;

pub use crate::ooa_2::get_subtypes;

pub use crate::ooa_2::get_supertype;

pub use crate::sarzak::SarzakModel;

pub use crate::drawing::UUID_NS as DRAWING_UUID_NS;

pub use template::emit_generated_code_comments;

pub use template::macros::emit_binary_main;

pub use template::macros::emit_many_unconditional;

pub use template::macros::emit_many_unconditional_lookup;

pub use template::macros::emit_one_conditional;

pub use template::macros::emit_one_conditional_lookup;

pub use template::macros::emit_one_unconditional;

pub use template::macros::emit_one_unconditional_lookup;

pub use template::types::emit_doc_string;

pub use template::types::emit_enum_main;

pub use template::types::emit_singleton;

pub use template::types::emit_struct_main;

Modules

config

Code Generation Configuration

context

Context for Extrusion

macros

Macros to make code generation nicer. 😉

template

types

Things that will help when generating types

Structs

DrawingObjectStore

JSFormat

SarzakObjectStore

Traits

Extrude

Extrude a new type

ReadModel

Read a Cuckoo Model file

WriteModel

WriteObjectStore