Function nut::codegen::template::types::emit_doc_string

pub fn emit_doc_string(input: &str) -> CachingContext

Generate struct/enum Documentation

The text from the tool is really long lines separated by \n. We split the lines up on unicode word boundaries and then reconstitute keeping the generated line length less than MAX_LEN characters.

It would be extra sweet to extract the doc links and construct pointers to known types. For example, “points at an Object”, would turn into “points at an [Object][o]”, and we’d generate an “[o]: nut::sarzak::Object” at the bottom of the comments.

This is still pretty cool compared to before. The long strings really got to me.