Module nut::drawing

The Drawing Domain

These are the bits that belong to Drawing, the domain of drawing boxen and lines!

Re-exports

pub use get_a_across_r3;

pub use get_bui_across_r7;

pub use get_bui_across_r8;

pub use get_r_bin_across_r12;

pub use get_many_edg_across_r14;

pub use get_oui_across_r1;

pub use get_p_across_r13;

pub use maybe_get_one_edg_across_r3;

Structs

Anchor

An anchor, or anchor point, is the location where an arrow from a relationship attached to an object.

BinaryUI

This represents additional information necessary to render a Binary relationship in the user interface.

IsaUI

This represents additional data necessary to render an Isa relationship in the user interface.

ObjectStore

ObjectUI

Render a rectangle

Point

A point is a two-tuple that represents a location on the drawing canvas.

Enums

Edge

An attachment point for an Anchor

RelationshipUI

Additional information necessary to render relationships in the user interface.

Constants

Bottom

The bottom of a rendered box

Left

The left side of a rendered box

Right

The right side of a rendered box

Top

The top edge of the rendered box

UUID_NS