Macro nut::get_subtypes

macro_rules! get_subtypes {
    ($obj:ident($rel:expr), $store:expr) => { ... };
    ($rel:expr, $store:expr) => { ... };
}

Macro to get the Supertype from an Isa relationship

We are given a reference to the store, a reference to an Object, and the name of the relationship to traverse. E.g.,

let animal = get_supertype!(cat("R1"), store);

We also support the form wherein a UUID is passed in:

    let animal = store.get_obj("Animal").unwrap();
    let r1 = animal.rels.get("R1").unwrap();
    let animals = get_subtypes!(r1.value, store);

I tried using a RelPointer instead of a UUID, but the macro won’t compile that way. I have no idea why, it’s just supposed to be a stupid macro.