Module nut::sarzak

Sarzak OOA of OOA

This surfaces the latest bit’s of sarzak, rather than having to delve deeply into one of the ooa_* crates.

I’m trying to be careful to not use re-exports from the ooa_* crates.

Re-exports

pub use get_obj_across_r17;

pub use get_r_from_across_r6;

Structs

Associative

Attribute

AttributeName

AttributeName NewType

Binary

A relationship between two objects

Isa

Object

The mighty Object

Referent

Referrer

RelPointer

A struct to store a pointer and type

SarzakModel

A model from the tool

Subtype

Supertype

This is a brand new class and has no From impl.

Enums

Cardinality

Conditionality

RelSide

An enum to disambiguate relationship type

Relationship

An association between two or more objects

Type

Constants

AS_IDENT

AS_TYPE

UUID_NS

Traits

ReadSarzakModel

WriteSarzakModel