Enum nut::sarzak::Relationship

pub enum Relationship {
    Binary(Uuid),
    Isa(Uuid),
    Associative(Uuid),
}

An association between two or more objects

A relationship connects two or more objects, depending on the type of relationship.

There are three:

Binary

Isa

Associative

A binary relationship connects two objects. An isa relationship connects two or more objects. One objects is the Supertype, and the others are Subtypes. An associative relationship associates three objects. TODO: I’ll explain once it’s actually implemented.

A relationship has a number, that is unique in the model.

Variants

Binary(Uuid)

Isa(Uuid)

Associative(Uuid)

Implementations

impl Relationship

pub fn get_number(&self, store: &ObjectStore) -> Option<u16>

pub fn get_id(&self) -> &Uuid